visualmemorysearch/backend/app/config/settings.py
"""
Configuration settings for the Visual Memory Search application.
"""

import os
from pathlib import Path
from typing import Optional
from pydantic_settings import BaseSettings
from pydantic import Field
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

class Settings(BaseSettings):
    """Application settings."""
    
    # Application settings
    app_name: str = "Visual Memory Search"
    app_version: str = "1.0.0"
    debug: bool = Field(default=False, env="DEBUG")
    
    # Server settings
    host: str = Field(default="0.0.0.0", env="HOST")
    port: int = Field(default=8000, env="PORT")
    
    # Screenshot directory
    screenshot_dir: str = Field(default="test_screenshots", env="SCREENSHOT_DIR")
    
    # OpenAI settings
    openai_api_key: Optional[str] = Field(default=None, env="OPENAI_API_KEY")
    openai_model: str = Field(default="gpt-3.5-turbo", env="OPENAI_MODEL")
    
    # Model settings
    embedding_model: str = Field(default="all-MiniLM-L6-v2", env="EMBEDDING_MODEL")
    vision_model: str = Field(default="microsoft/git-base", env="VISION_MODEL")
    
    # Search settings
    max_search_results: int = Field(default=10, env="MAX_SEARCH_RESULTS")
    similarity_threshold: float = Field(default=0.7, env="SIMILARITY_THRESHOLD")
    
    # File upload settings
    max_file_size: int = Field(default=16 * 1024 * 1024, env="MAX_FILE_SIZE")  # 16MB
    allowed_extensions: list = Field(default=["png", "jpg", "jpeg", "gif", "bmp"])
    
    # Logging
    log_level: str = Field(default="INFO", env="LOG_LEVEL")
    log_file: Optional[str] = Field(default=None, env="LOG_FILE")
    
    class Config:
        env_file = ".env"
        case_sensitive = False

# Global settings instance
_settings: Optional[Settings] = None

def get_settings() -> Settings:
    """Get the global settings instance."""
    global _settings
    if _settings is None:
        _settings = Settings()
    return _settings
