visualmemorysearch/backend/requirements.txt
opencv-python==4.9.0.80
pillow==10.2.0
pytesseract==0.3.10
transformers==4.38.2
torch==2.2.1
sentence-transformers==2.5.1
faiss-cpu==1.8.0
scikit-learn==1.4.0
openai==1.12.0
python-dotenv==1.0.1
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.2.1
gunicorn==21.2.0
numpy==1.26.4
opencv-python-headless==4.9.0.80
# Authentication dependencies
python-jose[cryptography]==3.3.0
passlib>=1.7.4
bcrypt==4.0.1
httpx==0.26.0
email-validator==2.2.0 